# contracts/generated/rust/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "description": "FastAPI-compatible validation envelope retained as a stable HTTP contract.",
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
...
      "UserInfoResponse": {
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ],
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "type": "object"
          },
          "input": {},
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array"
          },
          "msg": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Music contributors"
    },
    "description": "",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "music-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/sessions/{token_prefix}": {
      "delete": {
        "operationId": "revoke_session",
        "parameters": [
          {
            "description": "Unique session token prefix",
            "in": "path",
            "name": "token_prefix",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}